Move partner auto-approve onto defineJob.
